// resources/js/individual-components/player-components/tops/kingdom-tops/types/kingdom-leaderboard-props.ts
import TopsApiResponse from "../../shared/types/tops-api-response";

export default interface KingdomLeaderboardProps {
    leaderboard: TopsApiResponse | null;
    loading: boolean;
    errorMessage: string | null;
    period: string;
    metric: string;
    search: string;
    onPeriodChange: (value: string) => void;
    onMetricChange: (value: string) => void;
    onSearch: (value: string) => void;
}